mycatalog/requirements.txt
certifi==2018.1.18
Django==2.0.2
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
mysqlclient==1.3.12
PyJWT==1.5.3
pytz==2018.3
